* replaced the int ThreadInfo states with a ThreadInfo.State enum
* added SLEEPING / WOKEUP, as kind of an analogy to TIMEOUT_WAITING / TIMEDOUT. The reason for this is that Thread.getState() is supposed to return Thread.State.TIMED_WAITING if a thread is waiting.

This is very weak, since the javadoc for TIMED_WAITING says "..A thread that is waiting for another thread to perform an action for up to a specified waiting time is in this state", which is clearly not true for sleepers, but the javadoc also says a Thread.sleep() call puts it into this state. I assume this is an implementation artifact of the jdk that just bubbled up, and nobody is supposed to use Thread.getState() for anything serious anyways.

What makes it weak for JPF is that - as opposed to TIMEOUT_WAITING - there is no explicit event (like notify()) that ever puts a SLEEPING back into RUNNING, since: (a) we don't model time, and (b) the desktop JVM gives no guarantees for actual sleep times, so we don't even want to model time (huge effort to miss potential defects)

As a compromise, we currently treat SLEEPING / WOKEUP as runnable, set it for the corresponding transition, and then automatically reset to RUNNING when the CG is processed. This means non-sleeper threads see a sleeper, and listeners can see that a sleeper was woken up when we simulate the time out.

The solution needs more work though since it relies on ThreadChoiceFromSet, which just happens to be our only Thread CG (for now). Thread.currentThread().getState() also returns TIMED_WAITING, which is wrong