(1) added VM notifications for exceptionBailout() (popped frames w/o matching handler) and exceptionHandled(). In both cases, the relevant stackframes can be obtained with ti.getTopFrame(). This is mostly to simplify listeners that monitor exception handling.
NOTE -  listeners which were not derived from ListenerAdapter or PropertyListenerAdapter have to be extended accordingly. This is why you should always use the adapters.

(2) Changed ThreadInfo.executeInstruction() to do setPC() *after* the instructionExecuted() notification. The way it was before, createAndThrowException() from a listener did report the next instruction as the culprit. This required a change of the Instruction.isCompleted(ti) implementation, which is for instance essential when obtaining InvokeInstruction arguments, but its still better than having different patterns for pre- and post-execution createAndThrows. Besides, the isCompleted() is now streamlined and only in Instruction.

Its also better to update the pc after notification because this way listeners can use ThreadInfo.setNextPC().

NOTE - if instructionExecuted() listeners relied on topFrame().getPC() != vm.getLastInstruction() they have to be adapted

(3) added a Trace.removeLastOp() so that trace elements can contain dynamic ops like method invocations