updated IdleFilter docu, esp. the 'action' options (warn,break,prune,jump)
added a boolean cg.break_single_choice option. It is for people who think of JPF as a model checker and not a VM, and therefore expect state matching at all points where a CG *could* be created.

refactored the Verify peer accordingly. There was an awfull lot of redundancy anyways

also added this to DefaultSchedulerFactory.getRunnablesIfChoices(), but I'm not sure this is the right thing to do. While being consistent, it can cause a lot of additional states in single-threaded programs with a lot of synchronized method calls (e.g. from libs)