use absolute pathnames when including files, so that relative paths within the included file (jpf.properties, mostly) get resolved with respect to the property file location, not with respect to the dir JPF was started from. This should make use of ${project}/.. prefixes in jpf.properties obsolete, which in turn makes NetBeans happy again (it doesn't know about keyword expansion, which means jpf.properties couldn't be included in project.xml)
added "@include_if = ?key?path" and "@include_unless = ?key?path" pseudo properties, to allow conditional property file inclusion. Both key and value are expanded prior to lookup, but it is not recommended to get too fancy with the keys. The standard use is from SUT project application properties not knowing if a certain JPF extension is loaded or not:

  # this adds the required paths for jpf-numeric (if not already loaded via site.properties)
  @include_unless = ?jpf-numeric?../../../../jpf-numeric/jpf.properties

  # this sets the vm.insn_factory.class (which does not coexist with other extensions)
  @include_if = ?jpf-numeric?${jpf-numeric}/numeric.properties
  ...

modified JPFConfigException creation to include the 'config' context such an exception is raised from (i.e. the misconfigured file)