[1] the big change: peer method executions now have their own NativeMethodInfos and NativeStackFrames, i.e. we have MethodInfo and StackFrame class hierarchies. The NativeStackFrame instances only execute two new, synthetic insns: EXECUTENATIVE and NATIVERETURN. The reason for this is mostly twofold:
(1) to have symmetric invoke/return instructions, no matter if the implementation of a method is mji or not

(2) to solve the old problem of not having a scheduling point upon return from a synchronized native/MJI method. Having only one on the invoke inns isn't enough since the peer method can of course have thread-global effects

As a side effect, we now don't have to cache the invoke arguments anymore, since the caller  stackframe is kept intact until NATIVERETURN execution.

Another even more beneficial side effect is that we now have a place to store data from peer methods that have to do a (potentially open) number of roundtrips (think native HashMap). The NativeStackFrame operands and locals are not used for arguments and return values for the corresponding method, they are completely yours for use from within peer methods.

The downside of all this is that there already was a lot of code that explicitly handled/assumed the old non-symmetrical behavior. This code has to be changed to reap the benefits mentioned above, mostly for two things:

(a) post-exec listeners on InvokeInstructions have to be aware of that the peer method has not been executed yet. Listen on the EXECUTENATIVE instead.

(b) peer methods have to be aware of that the caller stackframe is *not* on top. To get the caller (e.g. to get locals etc.), use ThreadInfo.getLastNonSyntheticStackFrame()

This required a number of regression tests to be adapted.


[2] added a VMListener.choiceGeneratorRegistered() notification. This is not the same as the already existing choiceGeneratorSet(), because it happens at the end of a transition, which enables listeners to override and continue the current transition. The choiceGeneratorSet() is always at the beginning of the next transition, i.e. after state matching etc.

[3] in the wake of [1], the IntChoiceDependenSet (temporarily?) fell victim to an Int/DoubleChoiceFromSet modification. Successive calls to getNextChoice() from within the same transition should always return the same value (which was used to start the transition), i.e. resolving field/local values has to happen in the CG constructor. We also have that functionality in StackFrame (although incomplete), so IntSpec is obsolete.