added instrumentation for forward / backtrack consistency checks. Dont't trust the forward check too much though since that might lead into a property violation
the expensive checks are controlled by a JVM.CHECK_CONSISTENCY final static, i.e. they should not be compiled in unless this is set. We don't want to use class-wise assertion enabling for this since we do have assertions within transitions

turned Thread.setState() into a check point for lockRef consistency. !! lockREf is now ONLY set if we really are blocked or waiting, not if this is just a running lock contender that will be blocked once a thread acquires the lock in contention