refactored refTid into ElementInfo as a state component and updated Restorers accordingly. refTid changes its meaning and is now the last thread that accessed this element. The intention is to use it as a replacement for the ATTR_TSHARED.
fixed bug in ConstructorPeer, which wasn't accounting for the extra operand space it needs to store the object reference. This would not have been fatal if DynamicStackFrame (which is going away anyways) would have correctly grown the operand stack, which it didn't. Also added a simple regression test for it. Thanks Steve.

moved the String peer intern() method to the same mechanism that is used internally, i.e. Heap.newInternString(). However, this implementation is not guaranteed to be correct for HeuristicSearches (doesn't preserve identities). Intern Strings ultimately have to be kept in a SUT space hashset