added FixedBitSet interface and BitSet64 as additional implementor (changed BitSet256 accordingly). This is to make it more easy to change the set size. Also added bounds checks to both implementors and a BitSet64Test
replaced single int refTid in ElementInfo with FixedBitSet. This should avoid killing heap symmetry while being reasonably fast and not too storage intensive. With this, moved the old ElementInfo.isSchedulingRelevant() away from isShared()/ATTR_TSHARED and over to checkUpdatedSchedulingRelevance(), which also updates the refTid set and checks for live threads in the set.  Yields almost a 50% speedup in the state intensive RobotManager.

changed cg.threads.break_start default to 'true'. NOTE - this is now important for field race detection. If Thread.start() would be still a left-mover, a thread might terminate before a competing thread ever touches the objects in question. There is a tradeoff involved, but it seems to pan out nicely: in general, we want to avoid early state space branches, but so far we could only do this by being very conservative with the operations that break the leftmoving transition. This effectively seems to have caused larger state spaces - less on the threads, more on potentially shared fields. Now we go on without breaking transitions until we actually know an object has been referenced from different threads, which esp. could make a huge difference with static fields.

removed the old RaceDetector - nobody was using it anymore anyways