renamed & fixed the old CanonicalizingFilteringSerializer (now InvMapFilteringSerializer, the recursive loop was using fixed bounds that interestingly caused the RobotManager statespace to shrink dramatically while still finding the error).
Added a new implementation for the first one that is based on ElementInfo cache data and a ReferenceQueue. This is more memory efficient and scales better in terms of number of serialized objects. In fact, it is so good we can probably safely skip the SCAHeap implementation that uses a special segment for class init related allocations.