changed the FilteringSerializer to use block copy for StackFrames. Unfortunatly that wouldn't help for NamedFields since there are two bit masks to consider (filtered and isReference), and the required bitset iterations would be probably slower
renamed ReferenceVisitor to ReferenceProcessor and added ElementInfoProcessor interface - these are very generic interfaces that can be used in various places.

added a TopFrameSerializer that only serializes TopFrames and the objects referenced from there. This works spectacularly even in complex cases like the (jpf-awt) RobotManager, but both the TopFrameSerializer and the AdaptiveSerializer are under-approximators that can be tricked if there are thread interactions that don't use shared data.

polished the SimpleDot listener a bit, to show all backtracks and add a little more style to the edges