fixed ElementInfo.notifies() bug that was testing for a scheduling point where it just needed a ThreadChoiceGenerator
fixed inconsistency between StackFrame.equals() and hash() - thanks Nathan

changed MJIEnv exception handling to store object references, so that callers can create and initialize their own exception objects

added a 'canBeUncaught' field to the Throwable model, to avoid NoUncaughtExceptionsProperty violation on control exceptions such as ThreadDeath. Without it, we can't let ThreadInfo.throwException() process them

gave ThreadInfo the same attribute handling like ElementInfo etc. This is esp. useful for orthogonal thread state that cannot be mapped into the ThreadInfo.State enum, which in turn is crucial for synchronization, i.e. not easily extensible. Trying to get this into one state model would cause serious explosion. So far, the only sticky attribute is ATTRS_STOPPED, but there might be more in the future.

used all this to have a first stab at Thread.stop(). This does not change underlying stackframes or runnables, properly defers exception processing until the stopped thread is scheduled, and should coexist with the existing exception handling. However, there are rough edges and a lot more corner cases to check. The biggest problem is that it needs to execute the respective native method if it has to kill a thread that is executing in one. This relies on those native methods not to have side effects in their bottom halfs that run counter to Thread.stop() semantics (if there is such, other than empiricals). However, there currently is no easy way around that because ThreadInfo cannot know what lock reacquisition and such has to be done from the bottom half to get the Thread back into a consistent running state. We use the MJIEnv exception caching to process the exception upon return from the native method, but that also relies on that the bottom half does not blatantly overwrite already set MJIEnv exception objects in case it want to throw its own. It's funny that a topic that gives the host VM so much grief does so also for JPF, for different reasons but in the same place (natives).