added StringSetMatcher config to enable/disable assertions
vm.por.skip_local_sync is now standard, i.e. lock insns do check (and set) if an object is shared. This uses the same refTid like the field access, which might not be optimal, but there's a tradeoff keeping an additional bitset per object, and it is rather unlikely that we have shared field access but no contentious locking (or vice versa).

added SystemState.removeNextChoiceGenerator() to allow choiceGeneratorRegistered() listeners to override the CG. Changed SystemState.setNextChoiceGenerator() to return true if there is non-null nextCG (after notification), which has to be checked by CG creators to see if an insn should be reexecuted. Not all setNextChoiceGenerator() callers have been modified accordingly yet, so be careful.

refactored locking insns to check for shared-ness, and to throw exceptions in case the thread has to block but no CG can be obtained (which indicates an error in a choiceGeneratorRegistered() listener).