fix for System.arraycopy() peer, which should ignore null elements when type checking reference arrays - thanks Jan
fix for Array.getLength() peer, which should throw a NPE for a null argument - thanks Jan

added .vm.bytecode.BytecodeFactory, which is the advent of the new InstructionFactories. Still not complete yet

added argument ctors for a lot of Instruction classes (still not complete) to get rid of the old BCEL specific setPeer(..) initialization. To avoid breaking stuff I have to temporarily add a lot of default ctors too, but those will go away (together with setPeer) once we make the move - DON'T USE THEM EXCPLICITLY