added a cg.enable_atomic option (default true), which controls if Verify.begin/endAtomic() do anything. This can be used to easily switch on/off atomics without changing sources
added a log info for Thread.start() execution that does not result in a scheduling point, which could be caused by an atomic section, a listener removing CGs, or an explicit call of the native Thread.start() w/o re-execution. If Thread.start() is not a right mover, we might loose paths

added createBegin/EndAtomicCG() to the SchedulerPolicy interface. The DefaultSchedulerFactory turns the endAtomic() into a right mover, to make sure the atomic thread does not finish w/o a transition break, and therefore the startees never detect shared objects that were created by the starter (DiningPhil problem).

added a warning to RunAnt in case it tries to deduce the tools.jar location when running on a JRE

added support for multiple <project>.test_classpath elements to RunTest, and a warning showing the test_classpath contents in case the test class was not found