fixed bug in String.getBytes(charSetName) native peer, which was just returning the charSetName
added support for readonly URLConnections. This overrides sun.net.www.protocol.http.Handler, which uses JPF configuration to map URL patterns to URLConnection classes. Added a gov.nasa.jpf.CachedROHttpConnection example which can be configured to read data from a cacheDir, and only reaches out to the internet once to get this data in case it is not cached. For performance reasons, data is kept in memory once read. A sample configuration might look like this:

  http.cache_dir = tmp
  http.connection = http://*.dtd -- gov.nasa.jpf.CachedROHttpConnection

To monitor what is going on, set

  log.info=http

With this, the JDK1.6 xerces stuff starts to work. Added a .test.xerces.SAXParserTest regression test that is using example data from Checkstyle, and a cached version of the DTD this refers to (don't delete the strange looking http%... file in the test source dir or the regression test will make network connections). Note the humongous amount of instructions that are executed within this test

Added support for a tmp dir that is added to .hgingore, to have a place for temporary files that should not be added to the repo. The tmp dir is removed when doing an "bin/ant clean"