started effort to make gov.nasa.jpf.jvm as independent of a specific Instruction class hierarchy as possible. Added a gov.nasa.jpf.jvm.InstructionFactory interface so that external projects won't even have to subclass .bytecode.InstructionFactory.
This requires abstractions of some insn types. Started with ReturnInstruction, TableSwitchInstruction and LookupSwitchInstruction, but there are more left to be done. The idea is to use only these interfaces in .jvm so that we don't imply any concrete Instruction hierarchy

merged (some of) Ivan's String peer methods

added a ByteCodeReaderAdapter for selective bytecode parsing