BooleanChoiceGenerator fields should be protected
ad hoc fix of the SuspendResumeTest, which was falsely assuming that @Before/@After methods get executed within JPF. They only get executed outside, either from JUnit or from RunTest. JPF has no idea if it is running from within a TestJPF.verify..(), nor should it have one.

the Thread peer was creating two CGs per suspend/resume. I don't see why we need them in the top *and* the bottom half, so I removed the second one. Suspend/resume is only an approximation. In reality, this behaves not really accumulating (->ThreadData.suspendCount) but seems to sometimes produce deadlocks due to races witin suspend() calls for internal locks, which probably happens deep down in platform specific VM code. The important behavior to preserve faithfully is to not release locks upon suspension, it is Ok if we over-approximate on the safe side for repetitive suspension.