added ObjectList + test as a utility class to manage attributes in ChoiceGenerator, StackFrames, Fields and the like (so far its only used for ChoiceGenerator - the other ones still need to be converted)
fixed problem with interupted threads that can reacquire the lock and therefore are runnable at interrupt time. The trick is that the interruptee has to stay in the contender list (Monitor) in case another thread in the meantime grabs the lock. That in turn creates the opportunity for multiple Monitor.lockedThreads entries in case the InterruptedException handler has nothing better to do than to re-block (which is quite common). For that, we have to add a uniqueness check to Monitor.cloneWithLocked() (more specifically, to its add() method) - thanks Qingzhou!

fixed a problem with Types.getJNITypeCode(), which wasn't correctly handling the various formats of Class.forName(), thus causing the GenPeer to cerate wrong method names. Also removed some outdated stuff from GenPeer, but didn't test much - thanks Nastaran!

added a TimeModel interface with two implementations: SystemTime (just returning System.currentTimeMillis()), and ConstInsnPathTime, which returns something that is proportional to the number of instructions executed in the current path. For that, replaced ThreadInfo.isFirstStepInsn with a more useful 'executedInstructions', which keeps a counter that is reset upon entry into executeTransition(), and updated after each Instruction.execute() (means it changes values between executeInstruction() and instructionExecuted() notifications). The ThreadInfo.isFirstStepInsn() can still be used in both pre- and post-exec contexts. While the ConstInsnPathTime is nice and small, it seems instruction path length might be valuable enough to make this a named CG field (ConstInsnPathTime uses a CG attribute)

added a Verify.setCounter(id,val)