changed thread start (incl main) to use a direct call. This makes thread termination due to ThreadDeath or passed UncaughtException handlers much more simple, since we now have a guaranteed DIRECTCALLRETURN for each thread. In addition, it allows us to use normal sync mechanisms for synchronized runs() methods, although we still need a dynamic RUNSTART pseudo-insn so that we can tell the very first thread insn execution from a repeated one (isFirstStepInsn() always holds for the first insn in a thread)
removed the "canBeUncaught" field from Throwable since it is misleading. The only uncaught Throwable that would not be treated as a NoUncaughtExceptionProperty violation would be java.lang.ThreadDeath (via stop()), and that one we still have to hand over to a UncaughtExceptionHandler if we have one (its a common problem not rethrowing it and therefore making a thread unstoppable - making thread stopping even worse, but we want to catch this).

restructured the dreaded UncaughtHandler processing to only make it direct call a configured handler if its not the standard ThreadGroup, which in its uncaughtException() processing does the stack dump that for us constitutes a NoUncaughtExceptionProperty violation. Unfortunately this default uncaughtException() does more - it first delegates up the parent group chain, and also filters out the ThreadDeath Errors, so just moving this to the ThreadGroup peer is not simple either and causes a lot of directcall overhead for the very common case of not having an overridden UncaughtExceptionHandler. In addition, if we do report a NoUncaughtExceptionProperty violation, we want to preserve the whole stack for inspection, i.e. not unwind it. This is why the UncaughtHandler is executed in a direct call on top of where the exception occurred.

The behavior is controlled with two new options: 'vm.ignore_uncaught_handler' does exactly that - ignore any set UncaughtHandler and directly report a NoUncaughtExceptionProperty violation (most of these handlers are just used for different reporting, which we don't want to treat as 'handled'). There also is a 'vm.pass_uncaught_handler' option that tells JPF what to do when returning from a uncaughtException() direct call. If set to true, this is treated as normal thread termination w/o property violation. There is no "best" setting, it really depends on what the UncaughtHandlers (are supposed to) do.