added sun.misc.Unsafe direct buffer support to enable java.nio.Bits clinit (which uses it to determine platform endianess). Quite some effort for supporting a non-public class, and not yet complete because we need all the put/getter. It's using a JPF byte[] object to store/read buffer data. The platform byte order can be configured with vm.endianess (which defaults to the host platform order), but there are probably many places which aren't aware of it yet. This also required adding sun.misc.JavaNioAccess to the SharedSecrets. Also finally added the sun.misc.VM peer, but it's only a dummy.
removed the defunct IterativeContextBounding and replaced it with a somewhat ad hoc MinimizePreemption heuristic search. It mostly does the same by prioritizing the states with the smallest number of preemptions, which can be easily obtained by checking ThreadInfo changes over consecutive ThreadChoiceGenerators that are scheduling points, with the caveat that the previous thread has to be still in the choice set of the next CG (indicating the previous thread is still runnable). This included adding a few APIs to ThreadChoiceGenerator, JVM and SystemState

refactored classfile to make the classfile data source a public type (ClassFileContainer, replacing the old internal ClassPath.PathElement). We now store the container in the ClassInfo, so that we can subsequently check from where a class was loaded. This could be used for automatic peer lookup

the big little change: by slightly extending the ClassFileContainer derived JarContainer to include a path prefix, we can now specify directories within jars. This in turn finally allows single jar distributions of native and model classes. build.xml was modified to include the whole of jpf-classes.jar under a 'classes' subdir in jpf.jar, which can then be used by specifying the classpath as "jpf-core.classpath = build/jpf.jar/classes". Quite amazing how simple this became with our own class loading.