a rather big changeset to address problems with restoring ThreadList contents. There apparently were cases where ThreadInfo index wasn't consistent with the ThreadList index. This has to be fixed in the larger context of the non-reorganizing ThreadList (since we never removed ThreadInfos from the ThreadList, the latter one was just growing infinitely, which is bad in the bad case where SUTs use lots of short living threads).
The other general issue here were the relicts of implicit index/id - container assumptions. Leave alone the name, this was already solved for ElementInfos, but not for ThreadInfo. The 'index' fields are now replaced with appropriate names (id, objRef)

We still don't remove ThreadList entries in this change set because the ElementInfo.refTid bitset for reference tracking is still in place. This probably has to be replaced with a ThreadInfo object list and adapted hashes that abstract the order of list entries. A thread id bitset won't work because it means we would have to remove it from all live objects once a thread terminates, which can be very expensive. We can't leave the id in there because it might get reused for a subsequent thread, in which case sharedness checks on this id would fail