fixed the ErrorTraceGenerator listener on-the-go to be a proper PublisherExtension, and to be aware or cascaded CGs
fixed a problem with Integer$IntegerCache clinit, which in Java 7 relies on a Java7-only method of sun.misc.VM to dynamically initialize the upper cache bound. Since we can't abandon Java 6 yet, the most compatible fix was to intercept the IntegerCache clinit and initialize the cache ourselves from the native peer. We still have a fixed upper bound (as in Java 6), but at least we don't have a gazillion intstructions executed that are irrelevant for verification.

While ByteCache and CharacterCache have fixed upper bounds and hence would not be required, we also intercept their clinit for optimization reasons. There are probably still some other missing cache classes to do.