moved .externalToolBuilders/Ant\ Builder.launch into existing 'eclipse' dir and removed blank
added a 'vm.reuse_tid' option that controls if thread ids are reused along the same path (when thread objects get recycled). Default is true to play it safely with apps that sequentially create a lot of shortliving threads. Set to false if you rely on different thread ids, but be aware that tid != ThreadList index