added a DoubleChoiceFromList/Set pair similar to IntChoiceFromList/Set
added respective getChoices(), to find out what the current order is

added a supportsReordering() and reorder(Comparator<T>) to the ..List/Set CG classes. This is probably used most for scheduling choices, as some sort of a CG local poor man's HeuristicSearch.

added a CGReorderTest that shows how to use reorder/reverse