fixed wrong src/.. entries in run-JPF Eclipse launch config
added a test-JPF Eclipse launch config, to debug tests

NOTE - both can only be used as generic (external project) launch configs if you don't need sources, which pretty much renders this useless since the only justification for launch configs is debugging

tried to alleviate the 'no hg' problem of .version generation by refactoring it in its own task, which is skipped if there is no .hg found. Also used "searchpath=true", but it is not clear to me why this would yield any different result as executables are located by the OS anyways. Maybe it heps for the missing hg path problem on Windows