added a java.util.Random model class to get rid of the nagging problem of static initialization of the original class, which pulls in a lot of stuff we don't want. The old peer-only also didn't support a number of operations (esp. in cg.enumerate_random mode), was using a lot of fragile, duplicated code, and still was not compatible with the real Random (esp. not taking the seedUniqifier into account). The new peer delegates about everything to a singleton instance. Downside is that is requires a rather nasty native state restoration via Unsafe. All operations except nextGaussian() and nextBytes() now have cg.enumerate_random modes.
Added a jpfRunTerminated() method to the ConfigChangeListener interface, to better accommodate removal of listener instances, which have been notorious sources for memory leaks when using the same Config object for multiple JPF runs (esp. if they are inner class instances that keep the encapsulating object alive). Such listeners are now supposed to remove themselves from within the jpfRunTerminated() notification, which is ConcurrentModification safe. The old scheme required storing a reference of the listener in the client code, which was costly, error prone and not always feasible (e.g. if the client was a static-only native peer that is not aware of JPF termination).

  THIS REQUIRES EXTENSION PROJECT UPDATE (RunJPF/RunTest.jar)

Another fallout of the Random modelling was the need to support more basic choice types (Long, Float), which created a problem with ..List/Set CGs. While such choice containers should make use of generic bases to avoid code duplication, this previously was not possible because ChoiceGenerator and the type specializations such as IntChoiceGenerator were abstract classes. To support choice-type parameterized generic container class hierarchies, ChoiceGenerator had to be turned into a generic interface, builtin choice type specializations had to be turned into concrete interfaces. To factor out the common CG code, a new abstract generic ChoiceGeneratorBase class was added, which functions as the root for all existing CG subclasses. This looks like

  interface ChoiceGenerator<T>
     interface IntChoiceGenerator extends ChoiceGenerator<Integer>, ...
  abstract class ChoiceGeneratorBase<T> implements ChoiceGenerator<T> {common CG-attribute code}
    abstract class NumberChoiceFromList<T extends Number> extends ChoiceGeneratorBase<T> {common container code}
    class IntChoiceFromList extends NumberChoiceFromList<Integer> implements IntChoiceGenerator
      ...
    class LongChoiceFromList extends NumberChoiceFromList<Long> implements LongChoiceGenerator
    ...

This requires slightly more overhead when creating new CG types, namely specifying the concrete type interfaces and returning the proper type class object. However, this is inevitable if we want concrete choice type instanceof checks (that are used in a lot of places throughout the JPF universe) where Java's type erasure based generics otherwise shoot us in the foot.

  THIS MIGHT REQUIRE EXTENSION PROJECT MODIFICATION

but only if extensions introduce new choice types. Such modifications are also mechanical (see above), they should not involve anything with respect to choice computation.