cleaned up and fixed JPF_java_lang_reflect_Method.invoke, which was using boxed objects for native type conversion. This is one of the major sites  that use explicit argument initialization and hence has to be cleaned up before adding the ArgumentInitializer mechanism that abstracts different StackFrame organizations (Dalvik passes arguments in the trailing locals/registers)
rewrote the reg test for Method.invoke argument conversion, which was relying on TestJPF implementation behavior that is borderline. In general, tests should be SIMPLE - if it fails we should not have to think about if this is due to failure in the test subject or just the test logic itself. This is one of the few places where simplicity trumps minimizing redundancy.

merged patch for SimpleDot, which so far wasn't stateRestored aware - thanks Yu Lin

Eeek - cleaned up imports from Eclipse, and it touched almost every file. Well, get used to large change sets, there's more to come with the StackFrame / JVM refactor