added a 'vm.extra_startup_classes' option, which can be used to preload classes so that their static init doesn't happen during the search.
Added debug versions of the JenkinsStateSet and the CFSerializer, which store each new state through a JPFOutputStream in a state.<id> file. Default location is the 'tmp' dir. This is only a crude beginning, it certainly needs more testing and could also use a dedicated listener that just prints the result of each transition. It's nicely diffable though, i.e. together with a graphical diff viewer it makes state comparison quite easy.