added an AllocInstruction tag interface, to identify allocation instructions without referring to a specific instruction set. implementors are NEW, NEWARRAY and ANEWARRAY
changed MJI method filter to ignore methods with a 'final' modifier. Only "public static ... (MJIEnv,int,..)" methods that are NOT 'final' (which is pointless for static anyways) are considered to be MJI methods. Although this is close to a hack, this was required because there is an exported but non-MJI  method in the java.lang.reflect.Method peer that would otherwise be reported as orphan. This is considered to be rare but legit, so we need a way to mark as non-MJI

added first stab at a OOMEInjector listener and test (so far only location specs, but also with transitive scopes), to test SUT response to OutOfMemoryErrors. This should be completed with live object threshold and total heap memory threshold, although that is somewhat redundant with HeapTracker and far less useful to verify correct exception handling. This was motivated by the large object allocation problem in Android