.util.ClassSpec wasn't really a Spec, so it got renamed to TypeRef - that's what it is, just a way to name a class without having to load it, and without using a String
added a .util.TypeSpec, which really is a Spec to be used for config purposes (just like LocationSpec, FieldSpec and MethodSpec). It even uses the same base (FeatureSpec), but that should be probably refactored

used this to add type based OOME injection

added sun.misc.SharedSecrets.getJavaAWTSecret(), since it seems to be used in Java 1.6.0_31 on OS X (the security hotfix). The drag is its gone in Java 1.7, so we have to add a model of the interface. It is currently referenced from java.util.TimeZone, but luckily enough just in a null check