moved .bytecode/Instruction up into .vm so that .vm can add to/depend on it for execution counts (which are the basis for search global object references). The .bytecode classes are now derived from a common .bytecode.JVMInstruction. The package names should reflect this too (.jvm -> .vm, .bytecode -> .vm.jvm), but this is going to be an even bigger rename. However, we will do it for v7 so that we have a clear layer basis for Dalvik emulation (which becomes just a .vm.dalvik subdir)
  !! THIS REQUIRES EXTENSIONS TO BE ADAPTED (rename) !!

added some more util containers (SortedArrayObjectSet) that are/were used for the ExecutionCount, but this is still in flux

first stab at ExecutionCount, which is based on ClosedMemento restorers and uses SystemAttributes added to the relevant Instruction objects (NEW etc.). Since these objects capture not only the location (instruction object) but also the caller context (all parent stackframe PCs), this is a double nested structure insn->{thread->{cc,count}} - not a great solution.

--HG--
branch : v7-dev