slight mod of TestJPF to allow execution of all public arg-less instance methods, which comes in handy when adding benchmarks to TestJPF classes (should not be in JUnit test methods)
added Predicate- interface based block remove to ImmutableObjectTable. Our heap use case has a pattern like "{set|get}{remove}", i.e. removes just happens in blocks at the end of GC. Immutable tries aren't the most efficient data structures when it comes to removing elements since they have to modify the whole path to the removed element. The idea is to let the trie itself do the traversal and determine with a user supplied predicate which nodes/values have to be deleted. This way, we can do this recursively bottom-up and only have to allocate twice per modified node (the new elements array and then the node). Speedup is >4x, it would be even better (for consecutive key removes) if all nodes would store consecutive keys.

added a generic CountDown, which is really just a non-local goto to get out of complicated nested loops - if it counts down to 0 it throws itself

--HG--
branch : v7-dev