changed PersistentIntMap into an abstract base so that we can have protected abstract methods and inner types
all recursions now go through the PersistenMap so that they can be overridden. The nodes got demoted to relatively simple containers. This was required since staging would have required multiplying the Node hierarchy (with a new base to factorize) - a typical overlapping class hierarchy problem. The downside is a generic node element access (getElement()/setElement()), but since Node is not visible to clients anymore this is tolerable. Less tolerable is that we need to make the OneNode.nodeOrValue non-final to support merge node modification. Maybe it would be better to clone it, but on the other hand the other nodes are also modified (replacing an element) for the merge node case, so the 'final' isn't really doing it anyways.

PersistentStagingMsbIntMap now does get(), process() and set() with proper stageNode handling (incl. merge). The removes are still missing

--HG--
branch : v7-dev