some more PersistentIntMap docu and minor refactoring
added a snapshot mechanism to ObjVector, although it is not too general as it only breaks up the elements into chunks at gaps exceeding a certain size. It is a compromise between storage and performance, assuming that GC usually rips bigger holes into the heap, and bitmaped storage would not pay off for our typical heap size and (expected) fragmentation. Still needs to be validated once the SGOIDS are in place

added benchmarks with JPF specific usage mix of add/block-remove/clone. Well, PersistentIntMap performs favorably against HashMap, but not ObjVector. Typical case of better-is-the-enemy-of-good, or in this case simple-is-better

changed propertyViolated reporting to use Search.currentError instead of lastError. There are apparently listeners that reset the errors

--HG--
branch : v7-dev