moved randomizing config and policy enum into ChoiceGeneratorBase where it is used. SystemState only has to know that there is randomization in place. We still need CG.randomize() as a ctor since CGs can be directly created by clients (listeners etc) that should not have to be aware of randomization, and the randomization might require explicit enumeration of CGs that are otherwise dynamic (e.g. IntIntervalGenerator)
separated startup class registration and initialization to more explicitly deal with having to create StaticElementInfos before having class objects (for which we already need the startup classes to instantiate). During runtime, class registration first creates the class object so that its reference value is available for StaticElementInfo creation and does not have to be set after the fact. This comes in handy to initialize the referencingThreads set for SEIs, but the EI initialization sequence needs to be reworked.

recompute sharedness (which is a state stored attribute) in the EI ctor chain. The referencingThreads set might be global, hence resetting the attribute might otherwise result in inconsistent attributes/referencingThreads.

This also means we can't make the referencingThreads.add() optimization in EI.getInstanceWithupdatedSharedness() since we might get here after a backtrack. That is not ideal either. Maybe sharedness attribute handling should also go through the ThreadTrackingPolicy, but that would collide with explicit control.

added the requesting thread of requiresClinit() to the referencingThreads if there are clinit frames pushed. If not, subsequent exec should make use of either the class object or the type. If there are clinits, chances are the caller is going to re-execute before touching such insns, hence we should record the thread so that contenders get CGs on competing insns.

--HG--
branch : v7-dev