changed TestJPF behavior to not execute a test main() method from inside a verifyXX() but directly execute the enclosing method via target.entry. This seems more intuitive since we are already executing the method, and the best model is that the verifyXX() block is just executed under JPF. It seems counter-intuitive that it would branch to main(), which could actually execute different methods since main could bypass the standard TestJPF.runTestsOfThisClass(args). This also has the added benefit that we don't have to fiddle with argument extension or different targets in createJPF anymore - every JPF invocation now goes through runTestMethod(), which gets the test method from properties that are set in createJPF().
the low level TestJPF API (noPropertyViolation(), unhandledException() and the like) is a bit error prone if used for target methods other than the enclosing one. Use setTestMethod(cls,name) as the first argument if the enclosing method is not the test method.

Rewrote MJITest to conform to the standard JPF test infrastructure

rewrote ExceptionInjector to resolve exception types once the exceptions are injected. It previously resolved the types during vmInitialized(), but that is too late if the main class or any of the other startup classes should be instrumented (which happened during classLoaded(), which will precede vmInitialized()). Alternatively it could also have been at the beginning of the first classLoaded, since we obviously can resolve types at this point, but that would interfere with the startup registration and also does not deal with the fact that we need to create the exception objects through the right classloader (ExceptionInjector is meant to be a test tool for handlers, i.e. the exception types have to match).

--HG--
branch : v7-dev