Using id/uniqueId in ClassInfo.hashCode()/equals() violates the hashCode/equals invariant since the ids are not set before the ClassInfo is registered. It was just a more efficient shortcut for name/classLoader based hashing, but most cases that require it (registration within classLoaded() listeners) should use attributes nowadays anyways (more efficient, less memory, doesn't need additional reset).
This raises the bigger issue of ClassInfo initialization steps. So far, there are three:
- ClassInfo construction (just creates the ClassInfo and its supers)
- registration (creates StaticElementInfo, registers it in the classLoader.statics and creates companion java.lang.Class object)
- initialization (execution of clinits, if the class has one)
Registration enables references from the SUT, and before this is done the ids are not valid. If registration would be pulled into the construction phase, it would require to pass a ThreadInfo context through the (recursive) construction chain in ClassLoaderInfo/ClassInfo/ClassFileContainer, and it could lead to false positives since listener/peer internal ClassInfo requests would be linked to a specific thread. While the explicit registration puts more burden on the clients, separating between internal and SUT visible ClassInfo requests seems more valuable.

moved the classLoaded notification from ClassInfo ctors into register[System]Class(ti). This reflects the fact that listeners or peers might load ClassInfos that are not used in the SUT, and such loads should not be broadcasted.

--HG--
branch : v7-dev