fixed SingleProcessVM.createApplicationContext(), which was not properly setting SUT arguments
added a PersistentTidSet, which is now used from within OverlappingContenderPolicy. This does not modify existing ElementInfo.referencingThreads contents but clones the ElementInfo with a new PersistentTidSet upon add/remove of threads. While this puts more burden on callers of sharedness updaters (mostly via ElementInfo.getInstanceWithUpdatedSharedness(ti)), it ensures that sharedness depends only on the executed path and not the search history, which also means we can now store/replay paths that depend on sharedness to reproduce races etc.
To ease the caller burden a bit, there are now ThreadInfo.get[Modifiable]ElementInfoWithUpdatedSharedness(ref) methods, and for the sake of efficiency the caller should use most fitting ThreadInfo method to obtain the ElementInfo.
NOTE - this might change the number of states in some examples.