fixed missing executedInstruction update for hidden execution. If it isn't updated the isFirstStepInsn text fails
added a generic CheckEvent and corresponding evaluation in EventProducer as a means to specify properties in event trees

added an artificial NoEvent that is ignored in EventProducer.processNextEvent. This is handy if we need to create anchor
events before we get recursive (e.g. for permutation trees)

added a MJIEnv.getApplicationContext