added a .vm.CompoundChoiceGenerator if we want to collect choices from several (homogenous) CGs
added the dreaded interleave operation for Event trees. This is of course a combinatorial bomb and should only be used if the contributing trees produce very short traces since it is (sum1..N(ni))!/(prod1..N(ni!). Unfortunately it is the only sensible operation for combining event sequences of parallel systems (such as orthogonal statechart reasons) so it should be supported at the same level as other Event/tree constructors. The corresponding CG (SectionedEventGenerator) will probably be moved out of jpf-core though