fall back to 'junit.home' property for junit lookup, which can be either set explitly in local.properties (not a managed file) or by using a JUNIT_HOME environment var. That way, jpf-core does not depend on fiddling with the global CLASSPATH
re-instated the 'gov.nasa.jpf' prefix for '.class-pattern' specs in RunTest. We still need to refactor so that test classes are loaded by different class loaders in order to avoid carry over of static fields (should be optional)

fixed missing TestEventTree.isCompletelyCovered(), which prints out a warning if there was no 'expected' specification in the tree ctor set. Not very elegant, but at least avoids having to create an expected array in JPF executed code