fixed path replay bug in ChoicePoint/ChoiceSelector/ChoiceGeneratorBase, the most serious one being that ChoiceGeneratorBase.select() was not resetting prior to advance
added a ChoiceGenerator.getNumberOfParents() for quick path depth checks, added a VM.printChoiceGeneratorStack() for debugging

various fixes in EventTree construction, the most important one being that we need to add NoEvents to the tree in case we add a path that is an extension of an already existing one, and the various traversers like getPaths() have to be aware of if the client needs to know about NoEvents (e.g. in addPath) or if only visible events should be shown. This was creating wrong trees for embedded anyCombinations, and was generating way too many branches for interleaving

added the little NullTracker plus reg test, which is geared towards telling you where the null reference that causes an NPE came from. Shows how to tag values with attributes (even if there is no value)