started to concentrate sharedness and field lock protection into a new SharednessPolicy class, to finally have a clean separation and single configuration point for search global and path local policies
refactored FieldLockInfo into hierarchy that includes destructive overwrite and persistent
versions

fixed exposure bug that wasn't aware of BH execution due to other CGs (termination, interrupt).
This was causing the problems with the AQS. This is in general an unsolved problem - mix of
processing between TH and BH, and BH being unaware of in-between transitions. Used
StackFrame attribute for now, but this seems to warrant a different mechanism

refactored GET/PUT hierarchy to avoid the interface duplication. This is still something that
should be revisited once we move to Java 8 and have default interface methods. Factoring out
common get/put and instance/static are a clear case for multiple inheritance

fixed StackFrame getLongResultAttr() - thanks to Jakub Daniel