the big Scheduler overhaul. All scheduling now goes through a Scheduler master interface, which in turn is broken down into SyncPolicy (for all thread APIs, locks and signals) and SharednessPolicy (for access of shared objects). The interface to respective native methods and instructions now does not divide transition breaking methods into top and bottom half anymore, but allows to add new CGs upon re-execution, i.e. it supports empty transitions that exist solely to widen choices. This should better facilitate POR, especially since it combines all scheduling related functions in one component (Scheduler) instead of spreading it over instruction sets, peers and core constructs.
There is one generic DelegatingScheduler, which instantiates and then delegates to separate SyncPolicy and SharednessPolicy objects that can be configured

At this point there is only one generic AllRunnablesSyncPolicy, which first tries to locate runnables within the same ApplicationContext, and falls back to global runnables if none are found. This basically makes the old DistributedSchedulerFactory obsolete. Note that the default for lock release has changed and now also registers CGs. The old ContextBounding and PrioritySchedulerFactories still need to be re-implemented.

There is no functional change for the two SharednessPolicy implementations (Path and Global), but the interface has changed to make it more self explaining and conforming with SyncPolicy.

ThreadInfo got some methods to enable empty transitions, which should not happen silently. The facade for SharednessPolicy was removed since the Scheduler interface is to fat to be hidden.

Added a generic .util.InstructionState system attribute that can be used to store and retrieve execution state between (re-) executions of the same instruction. This is supposed to be used as a StackFrame attribute and can be a more efficient alternative for looking up the previous CG that was registered from the executing thread (in addition to provide a cache for state)

Fixed bug that caused <init> execution before <clinit>s of base classes were completed. This was caused by a non-recursive ClassInfo.isInitialized() guard around a recursive ClassInfo.initializeClass(). Since the latter one has all the required guards the first one is superfluous but was made recursive to avoid similar problems in extensions

Moved vm.max_transtion_length handling out of ThreadInfo into GOTO. This is a much better place since it always happens on a specific instruction type which should not confuse people as much as the previously used jump target instruction. It also doesn't have to check for termination corner cases. However, the current implementation does not handle nested loops too well. We should probably do some more analysis to choose a loop without non-filtered loop variables in order to help state matching (which is the reason to break)

Fixed false orphan peer method warnings. In order to allow peer method bindings from peer bases we used all peer methods for the reverse match, but only matched it with declared methods in the model class. We also have to detect if a potential orphan peer superclass method is already bound to a model superclass method