THIS STILL DOESN'T PASS CDL
modeling ThreadLocal, to become independent of complex containers using hashing and infrastructure such as Atomics and Unsafe. Assuming a small set of ThreadLocals, this uses a simple array that is updated through native methods. This finally also supports InheritableThreadLocals. The only downside is the bytecode init of inherited vars, which is required because of overridable InheritableThreadLocal.childValue() that would otherwiser require roundtrips. We might move it to native and do a method lookup though, similar to user defined ClassLoaders.

added a MJIEnv/ElementInfo.copyArray(), which is somewhat similar to System.arraycopy but doesn't support overlapping ranges and does only minimal type checking, so the user has to make sure it doesn't throw exceptions

--HG--
branch : java-8