first stab at java.util.concurrent.Exchanger. Timeouts are conceptually supported but still lack proper conversion. Zero timeouts also still need to be investigated since it is unclear how the waiter can immediately return without a TimeoutException
bad bug in AllRunnablesSyncPolicy in case the current thread is timeoutwaiting and the only runnable. This was falling back into the no-choice filter/singleton CG, which in turn caused backtracking on a new state

turned InstructionState into an abstract base. This saves an object per instance and would have been only useful to change the state for already registered attributes, which can also be done by subclassing.

fixed another bad bug in NativeMethodInfo, which was checking for round trips by doing identity comparison of the NativeStackFrame after the native exec returned. This fails spectacularly if the native method sets frame attributes, i.e. clones the frame. The resulting effect was a re-execution of the EXECUTENATIVE since it was treating the top frame as an overlay. This is now checked by using a StackFrame.originatesFrom(oldFrame)

--HG--
branch : java-8