implemented Exchanger timeouts. A timeout value of 0 means the first waiter immediately returns with a TimeoutException - this differs from Object.wait() timeout handling. Refactored Exchanger to do TimeUnit conversions for timeouts on the model side, to avoid duplicating code in the peer
Added filters (and related options) for potential left movers such as sleep(), yield(), interrupt and unpark (on running threads). Changed SyncPolicy.setsLockReleaseCG to include argument that indicates if the related operation actually caused threads to be unblocked, and use this argument in AllRunnablesSyncPolicy to filter cases where it does not.

Added type matcher for both sharedness and exposure management in GenericSharednessPolicy. Added callsite and type filter to setsExposureCG. Note that this refers to the type of the exposed object, not the field owner. There could be more filters but the best would be still to use the Sharedness/SyncPolicy interfaces to avoid exposures altogether.

changed the CGMonitor to show the insn that caused the CG, added options to only print upon setting the CG and to print choice values (upon advance). This is more useful to analyze the state space than grepping through ExecTracker output

--HG--
branch : java-8