fixed forgotten ClInitRequired check in DateFormat - thanks to Jiri Vahala
first stab at IntelliJ configuration. Note that the provided run configs require
the 3rd party ContextRun IntelliJ plugin
(https://github.com/m-lau/contextrun.git).

The mapping uses a IntelliJ "project" per JPF project, and a IntelliJ "module"
for each of the src/* subdirs (main, peers, classes,..). This is required to be
able to set the subdir specific output dirs (build/main,...), which are in turn
required to separate the jar contents and classpaths. We also keep the module
definitions (*.iml) files under .idea and share the run configs, so that we can
avoid any IDE content outside the IDE config dir.

Separating the projects has also the silver lining that simultaneously opened
projects are much better separated, which helps a lot if they contain the same
class names.

Note that the session specific .idea/workspace.xml is in .hgignore and hence
doesn't get cloned/merged

While this is all not the most IntelliJ compliant schema, the JPF directory layout
can stay the same and Eclipse/Netbeans compatibility should not be affected.